Adding coredump to logging
Summary:
Adding coredumps to logging

Why:
When we return logs from our containers we are using docker.api.container.logs. This method will only return messages in a container that were sent to STDERR or STDOUT. We are currently logging more data when there is a crash and we should expose that in our get_log function for our debug tool. In the core container for WhatsApp api, there is a crash log handler that will write the crash log to a log whenever there is an ungraceful exit. This diff will allow a user of the wadebug to pull any recent crash logs from the core container to help with debugging any recent crashes.

What I'm doing:
1. Create a method in docker_utils called get_core_dump_logs()
2. Pull all the latest files that were changed on the container.
3. Any that end in "-crash.log", call get_archive_from_container to get the contents of that file
4. Grab all the contents in all crash logs files and return that
5. Update get_logs method to call this new method get_core_dumps
6. Write the result to file

TODO: Coredumps is not properly configured on the Docker container. I noticed the coredump folder empty and also the crash log files empty as well. I need to explore more...

Reviewed By: thiagocmoraes

Differential Revision: D13394599

fbshipit-source-id: c500f2eb