Split get_log method in log_utils into separate methods - Refactor and Added unit tests
Summary:
Follow up diff. Based on comments made on this diff: https://our.intern.facebook.com/intern/diff/D12920351/

Splitting out log_utils.get_logs into seperate functions, each step in get_logs is split into its own function. Each function takes in a wa_container object and will try to grab the log it is written to grab. All new functions will return a string which is the file location of the log that is saved on the users local filesystem. For some log types ( i.e. coredumps ) they are specific to a certain type of container, so they will return a None value if the container is not of the right type.

Also added unit tests for each of these new functions that were added.

Reviewed By: mengyiyuan

Differential Revision: D13526016

fbshipit-source-id: 7921db60