pass phone number to server
Summary:
This diff adds support for populating the phone number field in all events sent to FB server.

There are now 3 different cases for the phone number field:
1. User's biz client has a registered number -> this number will be set in the event.
2. User's biz client is unregistered -> 'unregistered' will be set in the event.
3. Config file issues (no config file, bad credentials, etc.) -> phone number field will be 'null'

If in case (3) above, there will be __no prompt for the user__, as passing analytics back is best effort.  It may be too much to prompt them to fix just so we can pass analytics back.  This also means that from server side perspective, all we know is that the config was bad.  If we want to dive further, we can add logging for the tool and send back to server.

Reviewed By: thiagocmoraes

Differential Revision: D14597799

fbshipit-source-id: 81454162