Externalize dev mode toggle into an environment variable
Summary:
Previously we need to manually modify source file (`config.py`) to enable dev mode. This diff is to externalize this toggle to be an environment variable.

This toggle won't be part of the config file, because we don't want to confuse most of the customers with developer settings.

Also add visual indication on whether wadebug is currently running in dev mode.

Reviewed By: alanhhwong

Differential Revision: D14932768

fbshipit-source-id: 37f9cde8