Make check_mysql_version compatible with low PyMySQL version
Summary:
When PyMySQL version is low, like in the video, 0.6.3, it doesn't support cursor() function to be used in a context manager.

This diff use the connection in the context manager instead, which is compatible with the older version.

This diff also removes try...catch block, because the with context manager will take care of closing the connection.

Reviewed By: thiagocmoraes

Differential Revision: D15078366

fbshipit-source-id: 5b99bfb2