Get latest 3 hours of logs when running wadebug logs
Summary:
## Context
The current behavior of `wadebug logs` is to always get the last 10K lines of logs from each container. However, web and core containers may have different logging frequencies, and this behavior results in logs from different types of containers to be out of sync in timestamp.

This diff changes the behavior to always retrieve the latest 3 hours of logs instead.

Reviewed By: thiagocmoraes

Differential Revision: D19192650

fbshipit-source-id: aadfb3aa32e727434aecda5cf6a5a517abab3e1e