Make check_network action compatible with HA/MC setup
Summary:
### Original behavior:
Returns Error result when no Coreapp is running, because in a single instance setup, there should be one Coreapp running. This is confusing when users run checks from a host that is part of an HA/MC setup, where no Coreapp is running.

### This diff:
Skip the checks if no Coreapp container is running with clear messaging. Whether any Coreapp containers are running should not be the concern of this action (but the concern of check_containers action).

Reviewed By: alanhhwong

Differential Revision: D19487379

fbshipit-source-id: 4ee6e8be05d9af21b98b39e6d481e79ab4204e66