Change remediation step for wadebug errors
Summary:
### Context:

If a generic WADebug error occurs while running one or multiple actions due to issues like uploading results to FB, or it takes too long to generate logs, a `wadebug_error` will be thrown.

In such cases, we should ask the user to report the issue on Github rather than via Direct Support because it is not related to WaBiz setup.

Reviewed By: alanhhwong

Differential Revision: D19945462

fbshipit-source-id: 04f6f174846caad8e70fa16655f97bb7dc7465ba