Make containers_status action compatible with HA/MC setup
Summary:
### Context:
The current logic of containers_status covers the following checks:

1. Either core container or web container is missing - Error
2. The number of running coreapp containers is different from web containers - Error
3. No mysql container running - Error
4. More than one set of Web app and Core app are running. - Warning

This action is also marked as too complex by linter.

### This diff
In an HA/MC setup, none of these 4 criteria apply anymore. This diff changes the action to check the following:
1. If no WA container (including db) is running - Error
2. If some WA containers are stopped - Warning

The reason for 2 is that a BSP may have multiple sets of containers on each host, and some may be stopped due to reasons like the end client paused the integration, considering which I think a warning is more suitable.

Reviewed By: alanhhwong

Differential Revision: D20112195

fbshipit-source-id: da45686f5e3e6dbfdd89b03774e8f02cef2a7062